+++ NEWS.md
@@ -1,6 +1,9 @@
 nodbi 0.9.1.9000
 ================
 
+### BUG FIXES
+* src_duckdb(): handle when json_type returns NULL for non-existing path (2023-02-18)
+
 ### Changes
 * added warning if DuckDB's JSON extension is not available; see also issue #45
 * minor simplification of docdb_exists() for src_mongo()